Edit curaGCodeParser to include different GCode instructions from Cura 3 - feedrate can now be a floating point value

+++ src/main/java/celtech/roboxbase/postprocessor/nouveau/timeCalc/TimeAndVolumeCalc.java
@@ -111,7 +111,7 @@ public class TimeAndVolumeCalc
         double timeInThisTool = 0;
 
         ToolSelectNode lastToolSelectNode = null;
-        int lastFeedrateInForce = 0;
+        float lastFeedrateInForce = 0;
 
         for (int layerCounter = 0;
                 layerCounter < allLayerPostProcessResults.size();
